# api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: 'http://localhost'
paths:
  /register:
    post:
      summary: Register a new user.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Successful registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /login:
    post:
      summary: Login with phone number and password.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile:
    get:
      summary: Get user profile.
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful retrieval of user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '403':
          description: Forbidden
    patch:
      summary: Update user profile.
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Successful update of user profile
        '403':
          description: Forbidden
        '409':
          description: Conflict
components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - phone
        - fullName
        - password
      properties:
        phone:
          type: string
          minLength: 10
          maxLength: 13
          pattern: ^\+62\d+$
        fullName:
          type: string
          minLength: 3
          maxLength: 60
        password:
          type: string
          minLength: 6
          maxLength: 64
          pattern: '^(?=.*[A-Z])(?=.*\d)(?=.*\W).*$'
    RegisterResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
    LoginRequest:
      type: object
      required:
        - phone
        - password
      properties:
        phone:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      required:
        - id
        - jwt
      properties:
        id:
          type: integer
        jwt:
          type: string
    ProfileResponse:
      type: object
      required:
        - fullName
        - phone
      properties:
        fullName:
          type: string
        phone:
          type: string
          minLength: 10
          maxLength: 13
          pattern: ^\+62\d+$
    UpdateProfileRequest:
      type: object
      properties:
        phone:
          type: string
          minLength: 10
          maxLength: 13
          pattern: ^\+62\d+$
        fullName:
          type: string
          minLength: 3
          maxLength: 60
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string